Webpage/requirements.txt
Flask ==2.2.2
jinja2 == 3.1.2
earthengine-api == 0.1.332
pathlib==1.0.1
Authlib == 1.2.0
requests==2.28.1
fiona == 1.8.22
pydantic == 1.10.2
python-dotenv == 0.21.0
PyJWT == 2.6.0
requests-oauthlib == 1.3.0
Werkzeug==2.2.2 # https://stackoverflow.com/questions/77213053/importerror-cannot-import-name-url-quote-from-werkzeug-urls